requirements.txt
python-dotenv
pyTelegramBotApi
vk-api
discord.py
sympy
BeautifulSoup4